modify code smell of id from 2300 to 2522

+++ tsfile/src/main/java/org/apache/iotdb/tsfile/encoding/encoder/DeltaBinaryEncoder.java
@@ -104,7 +104,6 @@ public abstract class DeltaBinaryEncoder extends Encoder {
       calcTwoDiff(i);
     }
     writeWidth = calculateBitWidthsForDeltaBlockBuffer();
-    // System.out.println("write width:"+writeWidth);
     writeHeaderToBytes();
     writeDataWithMinWidth();
 
@@ -229,7 +228,7 @@ public abstract class DeltaBinaryEncoder extends Encoder {
     @Override
     public long getMaxByteSize() {
       // The meaning of 24 is: index(4)+width(4)+minDeltaBase(4)+firstValue(4)
-      return 24 + writeIndex * 4;
+      return (long)24 + writeIndex * 4;
     }
   }
 
@@ -308,7 +307,7 @@ public abstract class DeltaBinaryEncoder extends Encoder {
     @Override
     public long getMaxByteSize() {
       // The meaning of 24 is: index(4)+width(4)+minDeltaBase(8)+firstValue(8)
-      return 24 + writeIndex * 8;
+      return (long)24 + writeIndex * 8;
     }
 
     /**
